chore: convert dashboard and planning to server components
