+++ examples/coffee-warehouse-nextjs/app/warehouse/dashboard/loading.tsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+    return 'Loading dashboard...';
+}
