+++ examples/coffee-warehouse-nextjs/app/warehouse/loading.tsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+    return 'Loading...';
+}
