+++ examples/coffee-warehouse-nextjs/app/warehouse/page.jsx
@@ -1,4 +1,3 @@
-"use client";
 import DashBoard from "./dashboard/page";
 export default function Home() {
   return <DashBoard />;
