+++ examples/coffee-warehouse-nextjs/tsconfig.json
@@ -25,6 +25,6 @@
       ],
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "app/warehouse/dashboard/page.jsx", "app/warehouse/planning/page.jsx", "app/warehouse/info/page.jsx", "app/warehouse/profile/page.jsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "app/warehouse/dashboard/page.jsx", "app/warehouse/planning/page.jsx", "app/warehouse/info/page.jsx", "app/warehouse/profile/page.jsx", "app/warehouse/dashboard/dashboard-page.jsx"],
   "exclude": ["node_modules"]
 }
